+++ src/api/computed.ts
@@ -26,15 +26,11 @@ export const computedDecorator = createPropDecorator(
         // Optimization: faster on decorator target or instance? Assuming target
         // Optimization: find out if declaring on instance isn't just faster. (also makes the property descriptor simpler). But, more memory usage..
         const options = decoratorArgs[0] || {}
-        // <<<<<<< HEAD TODO: which one?
         asObservableObject(instance).addComputedProp(decoratorTarget, propertyName, {
-            ...options,
             get,
-            set
+            set,
+            ...options
         })
-        // =======
-        //         defineComputedProperty(instance, propertyName, { ...options, get, set })
-        // >>>>>>> master
     }
 )