+++ src/types/observableobject.ts
@@ -84,9 +84,13 @@ export class ObservableObjectAdministration
 
     read(owner: any, key: string) {
         // TODO: why was this relevant again?
-        if (owner !== this.target && owner !== this.proxy) {
+        if (
+            process.env.NODE_ENV === "production" &&
+            owner !== this.target &&
+            owner !== this.proxy
+        ) {
             this.illegalAccess(owner, key)
-            return
+            if (!this.values[key]) return undefined
         }
         // TODO: why was this relevant again?
         if (typeof key !== "string" || !this.values.hasOwnProperty(key)) return this.values[key] // might be on prototype
@@ -99,9 +103,8 @@ export class ObservableObjectAdministration
 
     write(owner: any, key: string, newValue) {
         const instance = this.target
-        if (instance !== owner && owner !== this.proxy) {
+        if (process.env.NODE_ENV === "production" && instance !== owner && owner !== this.proxy) {
             this.illegalAccess(owner, key)
-            return
         }
         const observable = this.values[key]
         if (observable instanceof ComputedValue) {
@@ -240,7 +243,7 @@ export class ObservableObjectAdministration
          *
          * When using decorate, the property will always be redeclared as own property on the actual instance
          */
-        return fail(
+        console.warn(
             `Property '${propName}' of '${owner}' was accessed through the prototype chain. Use 'decorate' instead to declare the prop or access it statically through it's owner`
         )
     }