+++ test/base/object-api.js
@@ -1,5 +1,17 @@
 const mobx = require("../../src/mobx")
-const { keys, when, set, remove, values, reaction, observable, extendObservable, has, get } = mobx
+const {
+    keys,
+    when,
+    set,
+    remove,
+    values,
+    entries,
+    reaction,
+    observable,
+    extendObservable,
+    has,
+    get
+} = mobx
 
 // TODO: duplicate test suite for not using utilities
 
@@ -68,6 +80,25 @@ test("object - set, remove, values are reactive", () => {
     expect(snapshots).toEqual([[3], [3, 4], [5, 4], [5]])
 })
 
+test("object - set, remove, entries are reactive", () => {
+    const todos = observable({})
+    const snapshots = []
+
+    reaction(() => entries(todos), entries => snapshots.push(entries))
+
+    expect(has(todos, "x")).toBe(false)
+    expect(get(todos, "x")).toBe(undefined)
+    set(todos, "x", 3)
+    expect(has(todos, "x")).toBe(true)
+    expect(get(todos, "x")).toBe(3)
+    remove(todos, "y")
+    set(todos, "z", 4)
+    set(todos, "x", 5)
+    remove(todos, "z")
+
+    expect(snapshots).toEqual([[["x", 3]], [["x", 3], ["z", 4]], [["x", 5], ["z", 4]], [["x", 5]]])
+})
+
 test("object - set, remove, keys are reactive", () => {
     const todos = observable({ a: 3 })
     const snapshots = []
@@ -101,6 +132,25 @@ test("map - set, remove, values are reactive", () => {
     remove(todos, "z")
 
     expect(snapshots).toEqual([[3], [3, 4], [5, 4], [5]])
+})
+
+test("map - set, remove, entries are reactive", () => {
+    const todos = observable.map({})
+    const snapshots = []
+
+    reaction(() => entries(todos), entries => snapshots.push(entries))
+
+    expect(has(todos, "x")).toBe(false)
+    expect(get(todos, "x")).toBe(undefined)
+    set(todos, "x", 3)
+    expect(has(todos, "x")).toBe(true)
+    expect(get(todos, "x")).toBe(3)
+    remove(todos, "y")
+    set(todos, "z", 4)
+    set(todos, "x", 5)
+    remove(todos, "z")
+
+    expect(snapshots).toEqual([[["x", 3]], [["x", 3], ["z", 4]], [["x", 5], ["z", 4]], [["x", 5]]])
 })
 
 test("map - set, remove, keys are reactive", () => {
@@ -144,6 +194,34 @@ test("array - set, remove, values are reactive", () => {
         [2, 3, undefined, 4],
         [2, 3, 4],
         [3, 4]
+    ])
+})
+
+test("array - set, remove, entries are reactive", () => {
+    const todos = observable.array()
+    const snapshots = []
+
+    reaction(() => entries(todos), entries => snapshots.push(entries))
+
+    expect(has(todos, 0)).toBe(false)
+    expect(get(todos, 0)).toBe(undefined)
+    set(todos, 0, 2)
+    expect(has(todos, 0)).toBe(true)
+    expect(get(todos, 0)).toBe(2)
+
+    set(todos, "1", 4)
+    set(todos, 3, 4)
+    set(todos, 1, 3)
+    remove(todos, 2)
+    remove(todos, "0")
+
+    expect(snapshots).toEqual([
+        [[0, 2]],
+        [[0, 2], [1, 4]],
+        [[0, 2], [1, 4], [2, undefined], [3, 4]],
+        [[0, 2], [1, 3], [2, undefined], [3, 4]],
+        [[0, 2], [1, 3], [2, 4]],
+        [[0, 3], [1, 4]]
     ])
 })
 
@@ -237,4 +315,5 @@ test("computed props are not considered part of collections", () => {
     expect(get(x, "y")).toBe(undefined) // disputable?
     expect(keys(x)).toEqual([])
     expect(values(x)).toEqual([])
+    expect(entries(x)).toEqual([])
 })